# docs/swagger.yaml
basePath: /
definitions:
  handlers.BalanceResponse:
    properties:
      current:
        type: number
      withdrawn:
        type: number
    type: object
  handlers.LoginRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  handlers.OrderResponse:
    properties:
      accrual:
        type: number
      number:
        type: string
      status:
        $ref: '#/definitions/storage.OrderStatus'
      uploaded_at:
        type: string
    type: object
  handlers.RegisterRequest:
    properties:
      login:
        type: string
      password:
        type: string
    type: object
  handlers.WithdrawRequest:
    properties:
      order:
        type: string
      sum:
        type: number
    type: object
  handlers.WithdrawalResponse:
    properties:
      order:
        type: string
      processed_at:
        type: string
      sum:
        type: number
    type: object
  storage.OrderStatus:
    enum:
    - NEW
    - PROCESSING
    - INVALID
    - PROCESSED
    type: string
    x-enum-varnames:
    - StatusNew
    - StatusProcessing
    - StatusInvalid
    - StatusProcessed
host: localhost:8081
info:
  contact: {}
  description: This is a sample server for Gophermart.
  title: Gophermart API
  version: "1.0"
paths:
  /api/user/balance:
    get:
      description: Get current balance and total withdrawn points
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/handlers.BalanceResponse'
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get user balance
      tags:
      - balance
  /api/user/balance/withdraw:
    post:
      consumes:
      - application/json
      description: Withdraw points from balance for a new order
      parameters:
      - description: Withdrawal
        in: body
        name: withdrawal
        required: true
        schema:
          $ref: '#/definitions/handlers.WithdrawRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Withdrawal successful
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "402":
          description: Insufficient funds
          schema:
            type: string
        "422":
          description: Invalid order number
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Withdraw points from balance
      tags:
      - balance
  /api/user/login:
    post:
      consumes:
      - application/json
      description: Login a user with login and password
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Invalid login or password
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Login a user
      tags:
      - user
  /api/user/orders:
    get:
      description: Get list of orders submitted by the user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.OrderResponse'
            type: array
        "204":
          description: No content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get list of orders
      tags:
      - order
    post:
      consumes:
      - text/plain
      description: Submit an order number for loyalty points calculation
      parameters:
      - description: Order Number
        in: body
        name: order
        required: true
        schema:
          type: string
      produces:
      - application/json
      responses:
        "200":
          description: Order already uploaded by this user
          schema:
            type: string
        "202":
          description: Order accepted for processing
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "409":
          description: Order number already exists
          schema:
            type: string
        "422":
          description: Invalid order number
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Submit an order number
      tags:
      - order
  /api/user/register:
    post:
      consumes:
      - application/json
      description: Register a new user with login and password
      parameters:
      - description: User
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/handlers.RegisterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT token
          schema:
            type: string
        "400":
          description: Invalid request
          schema:
            type: string
        "409":
          description: Login already exists
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      summary: Register a new user
      tags:
      - user
  /api/user/withdrawals:
    get:
      description: Get list of withdrawals made by the user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/handlers.WithdrawalResponse'
            type: array
        "204":
          description: No content
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal server error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Get list of withdrawals
      tags:
      - withdrawal
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"